backend/pyproject.toml
[project]
name = "learning-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "langchain>=0.3.18",
    "langchain-community>=0.3.17",
    "langchain-core>=0.3.34",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.10.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
